+++ src/index.ts
@@ -1,4 +1,0 @@
-
-export * from './lib/public_api';
-
-export * as sdk from './lib/sdk_api';