+++ src/lib/2.app/driving-ports/for-shared-externals-storage.port.ts
@@ -1,4 +1,4 @@
-import type { SharedExternal, SharedExternals, SharedVersion } from "lib/1.domain/public_api";
+import type { SharedExternal, SharedExternals, SharedVersion } from "lib/1.domain";
 import type { Optional } from "lib/utils/optional";
 
 export type ForSharedExternalsStorage = {