+++ src/lib/2.app/sdk_api.ts
@@ -1,2 +1,0 @@
-export * from './driving-ports/sdk_api';
-export * from './driver-ports/sdk_api';