+++ src/lib/4.config/import-map/use-default.ts
@@ -1,4 +1,4 @@
-import type { ImportMapConfig } from "lib/2.app/public_api"
+import type { ImportMapConfig } from "lib/2.app/config/import-map.contract"
 
 const useDefaultImportMap = (): ImportMapConfig => ({
     loadModuleFn: url => import(url),