+++ src/lib/6.mocks/adapters/manifest-provider.mock.ts
@@ -1,6 +1,6 @@
 import { ForProvidingManifest } from "lib/2.app/driving-ports/for-providing-manifest.port";
 import { MOCK_MANIFEST } from "../domain/manifest.mock";
-import { Manifest } from "lib/1.domain/public_api";
+import { Manifest } from "lib/1.domain";
 
 export const mockManifestProvider = ()
     : jest.Mocked<ForProvidingManifest> => ({