+++ src/lib/6.mocks/domain/manifest.mock.ts
@@ -1,4 +1,4 @@
-import { Manifest } from "lib/1.domain/public_api";
+import { Manifest } from "lib/1.domain";
 import { MOCK_REMOTE_ENTRY_SCOPE_I_URL, MOCK_REMOTE_ENTRY_SCOPE_II_URL } from "./remote-entry/remote-entry.mock";