+++ src/lib/index.ts
@@ -1,0 +1,9 @@
+export { initFederation } from './init-federation';
+export { NFError } from './native-federation.error';
+
+export * from './1.domain';
+export * from './2.app';
+
+export * as config from './4.config';
+
+export * as SDK from './sdk.index';