+++ src/lib/public_api.ts
@@ -1,7 +1,0 @@
-export { initFederation } from './init-federation';
-export { NFError } from './native-federation.error';
-
-export * from './1.domain/public_api';
-export * from './2.app/public_api';
-
-export * as config from './4.config';