+++ src/lib/sdk_api.ts
@@ -1,2 +1,0 @@
-export * from './2.app/sdk_api';
-export * from './create-nf-app';